Fix migration to be schema aware

+++ db/migrate/20250411114556_create_govuk_notifier_audit.rb
@@ -2,22 +2,26 @@
 
 Sequel.migration do
   up do
-    create_table Sequel[:govuk_notifier_audits].qualify(:public), id: :uuid do
-      primary_key :id
-      String :notification_uuid, null: false, unique: true
-      String :subject, null: false
-      String :body, null: false, text: true  
-      String :from_email, null: false
-      String :template_id, null: false
-      String :template_version, null: false
-      String :template_uri, null: false
-      String :notification_uri, null: false
-      DateTime :created_at
-      DateTime :updated_at
+    unless Sequel::Model.db.table_exists?(Sequel[:govuk_notifier_audits].qualify(:public))
+      create_table Sequel[:govuk_notifier_audits].qualify(:public), id: :uuid do
+        primary_key :id
+        String :notification_uuid, null: false, unique: true
+        String :subject, null: false
+        String :body, null: false, text: true  
+        String :from_email, null: false
+        String :template_id, null: false
+        String :template_version, null: false
+        String :template_uri, null: false
+        String :notification_uri, null: false
+        DateTime :created_at
+        DateTime :updated_at
+      end
     end
   end
 
   down do
-    drop_table Sequel[:govuk_notifier_audits].qualify(:public)
+    if Sequel::Model.db.table_exists?(Sequel[:govuk_notifier_audits].qualify(:public))
+      drop_table Sequel[:govuk_notifier_audits].qualify(:public)
+    end
   end
 end
